requirements.txt
flask==3.0.0
flask-socketio==5.3.6
kafka-python==2.0
fake-useragent==1.3.0